feat(example): add react navigation
