+++ example/package.json
@@ -9,8 +9,12 @@
     "start": "react-native start"
   },
   "dependencies": {
+    "@react-navigation/native": "6.0.11",
+    "@react-navigation/native-stack": "6.7.0",
     "react": "18.0.0",
-    "react-native": "0.69.1"
+    "react-native": "0.69.1",
+    "react-native-safe-area-context": "4.3.1",
+    "react-native-screens": "3.15.0"
   },
   "devDependencies": {
     "@babel/core": "7.18.5",
