+++ example/src/App.tsx
@@ -1,66 +1,16 @@
-import React, { useEffect, useCallback } from 'react';
-import { View, Platform } from 'react-native';
-import {
-  Event,
-  usePlayer,
-  PlayerView,
-  SourceType,
-} from 'bitmovin-player-react-native';
-import styles from './styles';
+import React from 'react';
+import { NavigationContainer } from '@react-navigation/native';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+import BasicPlayback from './screens/BasicPlayback';
 
-function prettyPrint(header: string, obj: any) {
-  console.log(header, JSON.stringify(obj, null, 2));
-}
+const Stack = createNativeStackNavigator();
 
 export default function App() {
-  const player = usePlayer();
-
-  useEffect(() => {
-    player.load({
-      url:
-        Platform.OS === 'ios'
-          ? 'https://bitmovin-a.akamaihd.net/content/MI201109210084_1/m3u8s/f08e80da-bf1d-4e3d-8899-f0f6155f6efa.m3u8'
-          : 'https://bitmovin-a.akamaihd.net/content/MI201109210084_1/mpds/f08e80da-bf1d-4e3d-8899-f0f6155f6efa.mpd',
-      type: Platform.OS === 'ios' ? SourceType.HLS : SourceType.DASH,
-      title: 'Art of Motion',
-      poster:
-        'https://bitmovin-a.akamaihd.net/content/MI201109210084_1/poster.jpg',
-    });
-  }, [player]);
-
-  const onReady = useCallback(
-    (event: Event) => {
-      prettyPrint(`EVENT [${event.name}]`, event);
-      player.play();
-    },
-    [player]
-  );
-
-  const onSourceLoaded = useCallback(
-    (event: Event) => {
-      prettyPrint(`EVENT [${event.name}]`, event);
-      player.getSource().then((source) => {
-        prettyPrint('SOURCE', source);
-      });
-    },
-    [player]
-  );
-
-  const onEvent = useCallback((event: Event) => {
-    prettyPrint(`EVENT [${event.name}]`, event);
-  }, []);
-
   return (
-    <View style={styles.container}>
-      <PlayerView
-        player={player}
-        style={styles.player}
-        onPlay={onEvent}
-        onPlaying={onEvent}
-        onPaused={onEvent}
-        onReady={onReady}
-        onSourceLoaded={onSourceLoaded}
-      />
-    </View>
+    <NavigationContainer>
+      <Stack.Navigator>
+        <Stack.Screen name="Basic Playback" component={BasicPlayback} />
+      </Stack.Navigator>
+    </NavigationContainer>
   );
 }
