+++ example/src/screens/BasicPlayback.tsx
@@ -1,0 +1,77 @@
+import React, { useEffect, useCallback } from 'react';
+import { View, Platform, StyleSheet } from 'react-native';
+import {
+  Event,
+  usePlayer,
+  PlayerView,
+  SourceType,
+} from 'bitmovin-player-react-native';
+
+function prettyPrint(header: string, obj: any) {
+  console.log(header, JSON.stringify(obj, null, 2));
+}
+
+export default function BasicPlayback() {
+  const player = usePlayer();
+
+  useEffect(() => {
+    player.load({
+      url:
+        Platform.OS === 'ios'
+          ? 'https://bitmovin-a.akamaihd.net/content/MI201109210084_1/m3u8s/f08e80da-bf1d-4e3d-8899-f0f6155f6efa.m3u8'
+          : 'https://bitmovin-a.akamaihd.net/content/MI201109210084_1/mpds/f08e80da-bf1d-4e3d-8899-f0f6155f6efa.mpd',
+      type: Platform.OS === 'ios' ? SourceType.HLS : SourceType.DASH,
+      title: 'Art of Motion',
+      poster:
+        'https://bitmovin-a.akamaihd.net/content/MI201109210084_1/poster.jpg',
+    });
+  }, [player]);
+
+  const onReady = useCallback(
+    (event: Event) => {
+      prettyPrint(`EVENT [${event.name}]`, event);
+      player.play();
+    },
+    [player]
+  );
+
+  const onSourceLoaded = useCallback(
+    (event: Event) => {
+      prettyPrint(`EVENT [${event.name}]`, event);
+      player.getSource().then((source) => {
+        prettyPrint('SOURCE', source);
+      });
+    },
+    [player]
+  );
+
+  const onEvent = useCallback((event: Event) => {
+    prettyPrint(`EVENT [${event.name}]`, event);
+  }, []);
+
+  return (
+    <View style={styles.container}>
+      <PlayerView
+        player={player}
+        style={styles.player}
+        onPlay={onEvent}
+        onPlaying={onEvent}
+        onPaused={onEvent}
+        onReady={onReady}
+        onSourceLoaded={onSourceLoaded}
+      />
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+    backgroundColor: 'black',
+  },
+  player: {
+    flex: 1,
+  },
+});
