--- example/src/styles.ts
@@ -1,12 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export default StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  player: {
-    flex: 1,
-  },
-});
