+++ package.json
@@ -48,8 +48,8 @@
     "@react-native-community/eslint-config": "3.0.2",
     "@types/jest": "26.0.24",
     "@types/lodash.omit": "4.5.0",
-    "@types/react": "17.0.39",
-    "@types/react-native": "0.67.7",
+    "@types/react": "18.0.15",
+    "@types/react-native": "0.69.2",
     "commitlint": "11.0.0",
     "eslint": "8.18.0",
     "eslint-config-prettier": "8.5.0",
